Add ClickHouse metrics bulk fetch to scatter-gather pipeline
reactionCount, commentCount, collectedCount come from ClickHouse, not PG.
The scatter-gather pipeline was hardcoding them to 0, breaking sort-by-
reactionCount. Now:

1. pg-sync load downloads metrics.csv from ClickHouse (TSV via HTTP)
2. scatter_gather loads it into HashMap<u32, (u32,u32,u32)> (~1.3 GB)
3. Gather phase looks up each slot's metrics for sort bitmap building

Also fixes cascading INT4/INT8 type errors in table_streams.rs from the
previous sync sidecar fix.

+++ src/pg_sync/table_streams.rs
@@ -302,7 +302,7 @@ pub(crate) async fn stream_tools(
         let mut current_images: Vec<u32> = Vec::with_capacity(100);
 
         for row in &rows {
-            if current_tool != Some(row.tool_id) {
+            if current_tool != Some(row.tool_id as i64) {
                 if let Some(tool_id) = current_tool {
                     flush_id_batch(
                         "toolIds",
@@ -312,7 +312,7 @@ pub(crate) async fn stream_tools(
                         &mut accum.filter_maps,
                     );
                 }
-                current_tool = Some(row.tool_id);
+                current_tool = Some(row.tool_id as i64);
                 current_images.clear();
             }
             current_images.push(row.image_id as u32);
@@ -377,7 +377,7 @@ pub(crate) async fn stream_techniques(
         let mut current_images: Vec<u32> = Vec::with_capacity(100);
 
         for row in &rows {
-            if current_tech != Some(row.technique_id) {
+            if current_tech != Some(row.technique_id as i64) {
                 if let Some(tech_id) = current_tech {
                     flush_id_batch(
                         "techniqueIds",
@@ -387,7 +387,7 @@ pub(crate) async fn stream_techniques(
                         &mut accum.filter_maps,
                     );
                 }
-                current_tech = Some(row.technique_id);
+                current_tech = Some(row.technique_id as i64);
                 current_images.clear();
             }
             current_images.push(row.image_id as u32);
